# lib/api-zod/src/generated/types/ramsAcknowledgeInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GroundworkOS API
 * OpenAPI spec version: 0.1.0
 */

export interface RamsAcknowledgeInput {
  name: string;
  role?: string;
  subcontractorId?: string;
}