# lib/api-zod/src/generated/types/ramsAttendee.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GroundworkOS API
 * OpenAPI spec version: 0.1.0
 */

export interface RamsAttendee {
  name: string;
  /** @nullable */
  role?: string | null;
  /** @nullable */
  subcontractorId?: string | null;
  acknowledged: boolean;
  /** @nullable */
  acknowledgedAt?: string | null;
}