# lib/api-zod/src/generated/types/ramsRecord.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GroundworkOS API
 * OpenAPI spec version: 0.1.0
 */
import type { RamsAttendee } from './ramsAttendee';
import type { RamsHazard } from './ramsHazard';

export interface RamsRecord {
  id: string;
  /** @nullable */
  jobId?: string | null;
  /** @nullable */
  jobTitle?: string | null;
  title: string;
  activity: string;
  riskLevel: string;
  status: string;
  hazards: RamsHazard[];
  ppe: string[];
  /** @nullable */
  briefedAt?: string | null;
  /** @nullable */
  briefedBy?: string | null;
  attendees: RamsAttendee[];
  /** @nullable */
  reviewDate?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}