# lib/api-zod/src/generated/types/ramsUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GroundworkOS API
 * OpenAPI spec version: 0.1.0
 */
import type { RamsHazard } from './ramsHazard';
import type { RamsUpdateAttendeesItem } from './ramsUpdateAttendeesItem';

export interface RamsUpdate {
  jobId?: string;
  title?: string;
  activity?: string;
  riskLevel?: string;
  status?: string;
  hazards?: RamsHazard[];
  ppe?: string[];
  attendees?: RamsUpdateAttendeesItem[];
  reviewDate?: string;
  notes?: string;
}